fix: make resize fire only once
