+++ packages/leva/src/hooks/index.ts
@@ -1,4 +1,4 @@
-export * from './useValue'
+export * from './useComputeInputSetters'
 export * from './useDeepMemo'
 export * from './useShallowMemo'
 export * from './useCompareMemoize'
@@ -11,3 +11,4 @@ export * from './useVisiblePaths'
 export * from './useValuesForPath'
 export * from './useInput'
 export * from './usePopin'
+export * from './useValue'
